pyproject.toml
[project]
name = "txt2sd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "coqui-tts[codec]>=0.27.5",
    "edge-tts>=7.2.7",
    "gtts>=2.5.4",
    "pdf2image>=1.17.0",
    "pdfplumber>=0.11.9",
    "pillow>=12.1.0",
    "torch>=2.10.0",
    "torchaudio>=2.10.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

